requirements.txt
fastapi>=0.116.0
openai>=1.99.0,<2.0.0
pydantic-settings>=2.10.0
pypdf>=6.0.0
python-multipart>=0.0.20
uvicorn[standard]>=0.35.0
